+++ src/main/java/net/tetro48/classicaddon/mixin/blocks/BlockMixin.java
@@ -1,5 +1,6 @@
 package net.tetro48.classicaddon.mixin.blocks;
 
+import btw.block.BTWBlocks;
 import net.minecraft.src.Block;
 import net.minecraft.src.BlockFurnace;
 import net.minecraft.src.EntityPlayer;
@@ -11,6 +12,10 @@ import org.spongepowered.asm.mixin.injection.Redirect;
 public abstract class BlockMixin {
 	@Redirect(method = "<clinit>", at = @At(value = "INVOKE", ordinal = 0, target = "Lnet/minecraft/src/BlockFurnace;hideFromEMI()Lnet/minecraft/src/Block;"))
 	private static Block notHideFurnaceFromEMI(BlockFurnace instance) {
+		return instance;
+	}
+	@Redirect(method = "<clinit>", at = @At(value = "INVOKE", ordinal = 50, target = "Lnet/minecraft/src/Block;hideFromEMI()Lnet/minecraft/src/Block;"))
+	private static Block notHideHopperFromEMI(Block instance) {
 		return instance;
 	}
 	@Redirect(method = "getPlayerRelativeBlockHardness", at = @At(value = "INVOKE", target = "Lnet/minecraft/src/EntityPlayer;isCurrentToolEffectiveOnBlock(Lnet/minecraft/src/Block;III)Z"))